+++ header/monte_carlo.h
@@ -7,9 +7,8 @@
 
 #ifndef MONTE_CARLO_H
 #define MONTE_CARLO_H
-
-void experiencesMonteCarlo(const int nb_experiences, const int nb_iterations, char use_random_number_already_generated) ;
-double monteCarlo(const int begin, const int end, char use_random_number_already_generated) ;
+void experiencesMonteCarlo(const long long int nb_experiences, const long long int nb_iterations, char use_random_number_already_generated);
+double monteCarlo(const long long int begin, const long long int end, char use_random_number_already_generated) ;
 int Question7() ;
 void calcTime(char use_random_number_already_generated) ;
 #endif
